requirements.txt
pdf2image
Pillow
streamlit
doctr
pandas
matplotlib
transformers
